refactor: add function for loading safe info
